requirements.txt
streamlit
opencv-python-headless
numpy
imutils
scikit-image
Pillow
